Add Interpolator and fix a bunch of nagging IDE messages
Some properties are now (re)initialized when a specific animation stage starts

+++ OpenVRNotificationPipe/Notification/Payload.cs
@@ -38,7 +38,7 @@ namespace OpenVRNotificationPipe
             public float distance = 0; // Translational offset
             public float horizontal = 0; // Translational offset
             public float spin = 0;
-            public float interpolation = 0; // Interpolation type
+            public int interpolation = 0; // Interpolation type
             public int duration = 100;
         }
     }
